Fix compiler error: -1073741819
Yeah, the compiler complains about goto but decided to say nothing on the logs WHATSOEVER. So we think this might due to SDK or NuGet update, but NO. The error was caused by this method where it calls goto statement to jump branch. used to invalidate the pointer.

WHAT
IN
THE
ACTUAL
FUCK
???